replace old 'skip-model-map-stringify?' option with a new global option
as well, by default now, keys will NOT be automatically stringified. the
user must turn this behaviour on explicitly. this is to reduce the level
of "magic" that happens behind the scenes (if there was a 100% fool-
proof method of stringifying/keywordizing back and forth as needed
without ever being an issue i would turn this behaviour on all the time)